Moved tests into own files/folders to prevent merge conflicts
