+++ tests/sources/test_cv.py
@@ -1,0 +1,17 @@
+from pashehnet.sensors.sources import ConstantValueSource
+
+
+class TestConstantValueSource:
+    """
+    Unit tests for ConstantValueSource class
+    """
+
+    def test_source(self):
+        """
+        Test that expected value is returned in iterative requests
+        :return:
+        """
+        val = 42
+        src = ConstantValueSource(val)
+        values = [next(src) for i in range(10)]
+        assert all(v == val for v in values)
